requirements.txt
# Core web and API
fastapi==0.115.12
uvicorn[standard]==0.34.2
requests==2.32.5
requests-toolbelt==1.0.0
python-dotenv==1.2.1
SQLAlchemy==2.0.49
greenlet==3.2.5
redis==6.4.0

# Frontend
streamlit==1.50.0

# LLM and orchestration
langchain==0.3.28
langchain-community==0.3.31
langchain-core==0.3.84
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.11
langsmith==0.4.37
openai==2.31.0

# Retrieval and NLP
faiss-cpu==1.13.0
jieba==0.42.1
rank-bm25==0.2.2
sentence-transformers==5.1.2
sentencepiece==0.2.1
transformers==4.57.6
tokenizers==0.22.2
huggingface_hub==0.36.2

# Medical imaging and ML
monai==1.5.2
nnunet==1.7.1
simpleitk==2.5.3
pydicom==2.4.4
dicom2nifti==2.6.0
nibabel==5.3.3
MedPy==0.5.2
connected-components-3d==3.26.1
scikit-image==0.24.0
scikit-learn==1.6.1
scipy==1.13.1
numpy==1.26.4
pandas==2.3.3
matplotlib==3.9.4
seaborn==0.13.2
opencv-python==4.13.0.92
Pillow==11.3.0
ImageIO==2.37.2
imagecodecs==2024.12.30
tifffile==2024.8.30

# PyTorch stack
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
accelerate==1.10.1
auto_gptq==0.7.1+cu118
peft==0.10.0
safetensors==0.7.0
optimum==2.1.0

# Model and data utilities
datasets==4.5.0
joblib==1.5.3
protobuf==6.33.6
PyYAML==6.0.3
regex==2026.1.15
tqdm==4.67.3
xxhash==3.6.0
filelock==3.19.1
fsspec==2025.10.0

# Document and parsing
PyMuPDF==1.26.5
Jinja2==3.1.6
MarkupSafe==3.0.3
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.20.0

# Scientific and math
contourpy==1.3.0
cycler==0.12.1
fonttools==4.60.2
kiwisolver==1.4.7
mpmath==1.3.0
networkx==3.2.1
sympy==1.13.1
pyparsing==3.3.2

# Async and HTTP
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
anyio==4.12.1
async-timeout==4.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
multidict==6.7.1
sniffio==1.3.1
yarl==1.22.0

# Pydantic and schema
annotated-doc==0.0.4
annotated-types==0.7.0
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
typing-inspection==0.4.2
typing_extensions==4.15.0

# Utilities
attrs==26.1.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.1.8
colorama==0.4.6
dill==0.4.0
distro==1.9.0
future==1.0.0
humanfriendly==10.0
idna==3.11
importlib_resources==6.5.2
lazy-loader==0.5
mypy_extensions==1.1.0
orjson==3.11.5
packaging==25.0
propcache==0.4.1
psutil==7.2.2
py-cpuinfo==9.0.0
pyarrow==21.0.0
pyreadline3==3.5.4
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2026.1.post1
referencing==0.36.2
rpds-py==0.27.1
shellingham==1.5.4
six==1.17.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tzdata==2026.1
urllib3==2.6.3
watchdog==6.0.0
zipp==3.23.0
zstandard==0.25.0

# JSON and validation
dataclasses-json==0.6.7
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
marshmallow==3.26.2

# Visualization and app helpers
altair==5.5.0
narwhals==2.19.0
pydeck==0.9.1
rich==15.0.0
tornado==6.5.5

# Training / model ecosystem extras
acvl_utils==0.2.1
batchgenerators==0.25.1
batchgeneratorsv2==0.3.2
blosc2==2.5.1
dynamic_network_architectures==0.3.1
einops==0.8.2
flatbuffers==25.12.19
gekko==1.3.2
graphviz==0.21
hf-xet==1.4.3
jiter==0.14.0
linecache2==1.0.0
ml_dtypes==0.5.4
modelscope==1.35.4
msgpack==1.1.2
multiprocess==0.70.18
ndindex==1.10.1
onnx==1.19.1
onnxruntime==1.19.2
rouge==1.0.1
smmap==5.0.3
traceback2==1.4.0
typer==0.23.2
typing-inspect==0.9.0
unittest2==1.1.0
uuid_utils==0.14.1
yacs==0.1.8

# Git and repository helpers
GitPython==3.1.46
gitdb==4.0.12

# Segment Anything dependency
segment_anything @ git+https://github.com/facebookresearch/segment-anything.git@dca509fe793f601edb92606367a655c15ac00fdf
